Move combinators (fmap.h, bind.h, foreach.h) to combinators/
