+++ include/absent/absent.h
@@ -1,8 +1,8 @@
 #ifndef RVARAGO_ABSENT_ABSENT_H
 #define RVARAGO_ABSENT_ABSENT_H
 
-#include "absent/bind.h"
-#include "absent/fmap.h"
-#include "absent/foreach.h"
+#include "absent/combinators/bind.h"
+#include "absent/combinators/fmap.h"
+#include "absent/combinators/foreach.h"
 
 #endif //RVARAGO_ABSENT_ABSENT_H
