+++ include/absent/combinators/foreach.h
@@ -1,7 +1,7 @@
 #ifndef RVARAGO_ABSENT_FOREACH_H
 #define RVARAGO_ABSENT_FOREACH_H
 
-#include "nullable.h"
+#include "absent/nullable.h"
 
 namespace rvarago::absent {
