+++ tests/fmap_test.cpp
@@ -1,4 +1,4 @@
-#include <absent/fmap.h>
+#include <absent/combinators/fmap.h>
 
 #include <optional>
 #include <string>
